requirements.txt
django==1.6.5
gunicorn
djangular
mongoengine
django_sse
redis
sse
gevent-socketio
greenlet
gevent 